Heaps of progress on WebView, mainly in context
Context is the module and ability to add callback methods for things
that occur in other processes and sandboxs (eg. js context
initialization). So far is just a POC and lot's of build changes to get
it working but am happy with how it's structured so can flesh out now.